requirements.txt
asgiref==3.6.0
autoflake==2.0.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-filter==22.1
django-crispy-forms==1.14.0
djhtml==1.5.2
gunicorn==20.1.0
idna==3.4
isort==5.11.4
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
django-ckeditor==6.5.1
Pillow==9.4.0
tzdata==2022.7